x86/entry/common: Split irq tracing in prepare_exit_to_usermode()
As in entry from user mode, lockdep and tracing have different
requirements. lockdep needs to know about the interrupts off state accross
the call to user_enter_irqsoff() but tracing is unsafe after the call.

Split it up and tell the tracer that interrupts are going to be enabled
before calling user_enter_irqsoff() and tell lockdep afterwards.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>