genirq: Reexport irq_to_desc() for PPC KVM
Commit f07147b162a1 ("genirq: Remove export of irq_to_desc()") breaks the
PPC - CONFIG_KVM_BOOK3S_64_HV=n build because the analysis of irq_to_desc()
usage missed that the creative fiddling in arch/powerpc/kvm/ is actually in
a module and not in built in code. The only real purpose is to access
irq_desc::kstat_irqs which can be solved differently, but not without some
surgery.

Reexport it when KVM_BOOK3S_64_HV is enabled. That means that all other
modular code especially drivers/* cannot rely on it, which was the whole
point of the exercise.

Fixes: f07147b162a1 ("genirq: Remove export of irq_to_desc()")
Reported-by: Ingo Molnar <mingo@kernel.org>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: Michael Ellerman <mpe@ellerman.id.au>